Use Int64 for byte counts

+++ Deluge/Sources/Deluge/Torrent.swift
@@ -21,9 +21,9 @@ public struct Torrent {
     /// The date the torrent was added to the server.
     public let dateAdded: Date
     /// The torrent's current download rate in bytes/s.
-    public let downloadRate: Int
+    public let downloadRate: Int64
     /// The torrent's current upload rate in bytes/s.
-    public let uploadRate: Int
+    public let uploadRate: Int64
     /// The torrent's current ETA.
     public let eta: TimeInterval
     /// The torrent's current progress.
@@ -52,8 +52,8 @@ public struct Torrent {
         name: String,
         state: State,
         dateAdded: Date,
-        downloadRate: Int,
-        uploadRate: Int,
+        downloadRate: Int64,
+        uploadRate: Int64,
         eta: TimeInterval,
         progress: Float,
         downloaded: Int64,
@@ -91,8 +91,8 @@ extension Torrent {
         guard let name = dictionary["name"] as? String,
             let stateString = dictionary["state"] as? String,
             let timeAdded = dictionary["time_added"] as? TimeInterval,
-            let downloadRate = dictionary["download_payload_rate"] as? Int,
-            let uploadRate = dictionary["upload_payload_rate"] as? Int,
+            let downloadRate = dictionary["download_payload_rate"] as? Int64,
+            let uploadRate = dictionary["upload_payload_rate"] as? Int64,
             let eta = dictionary["eta"] as? TimeInterval,
             let progress = dictionary["progress"] as? Float,
             let downloaded = dictionary["total_done"] as? Int64,
